refactor: extract shared LogViewer from logs page and logs sheet
The full-page logs route and the containers logs sheet duplicated ~90%
of their log-viewing logic (search, filtering, pinning, multi-select,
keyboard navigation, virtualization, highlighting, copy/download).
Extract the shared UI and logic into
src/features/containers/components/log-viewer/:

- LogViewer component owning the toolbar (page and sheet layout
  variants), virtualized list, search navigation, level filters, pins,
  selection, and download/copy actions
- use-log-search: search parsing, match computation, highlight
  rendering (pure parseSearch/computeSearchMatches + hooks)
- use-log-filtering: level/exclude filtering with filtered-to-original
  index mapping (pure computeFilteredLogItems + hook)
- use-log-pins: pin state, ring-buffer drop shifting, and wrap-around
  pin navigation (pure navigatePins/shiftPinsForDroppedLines + hook)

Behavior unification:
- structured log continuation grouping (groupRelatedLogEntries) is now
  enabled on the full-page route too, matching the sheet (fc1231a)
- the more accessible control variants win where the two files had
  diverged (aria-label/aria-pressed on toggles and icon buttons)
- the keydown listener is attached once and reads the latest handlers
  through a ref instead of re-subscribing on every callback change

The page keeps its recreate flow via a LogViewerHandle imperative ref;
the sheet remounts LogViewer per container via key, which stops the
stream and refetches exactly as before.

Adds vitest coverage for the extracted pure logic: filtered-index
mapping, search-match computation (regex, plain text, invalid regex),
and pin navigation.
