+++ frontend/src/routes/containers/$containerId/logs.tsx
@@ -1,73 +1,17 @@
 import { useQuery, useQueryClient } from "@tanstack/react-query";
 import { createFileRoute, useNavigate } from "@tanstack/react-router";
-import { useVirtualizer } from "@tanstack/react-virtual";
-import {
-	ArrowDownIcon,
-	ArrowDownToLineIcon,
-	ArrowLeftIcon,
-	CheckIcon,
-	ChevronDownIcon,
-	ChevronLeftIcon,
-	ChevronRightIcon,
-	CopyIcon,
-	DownloadIcon,
-	PauseIcon,
-	PlayIcon,
-	RefreshCcwIcon,
-	SearchIcon,
-	SquareIcon,
-	TerminalIcon,
-} from "lucide-react";
-import {
-	useCallback,
-	useEffect,
-	useId,
-	useMemo,
-	useRef,
-	useState,
-} from "react";
-import { toast } from "sonner";
+import { ArrowLeftIcon, ChevronDownIcon, TerminalIcon } from "lucide-react";
+import { useRef, useState } from "react";
 import { ThemeToggle } from "@/components/theme-toggle";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import {
-	DropdownMenu,
-	DropdownMenuContent,
-	DropdownMenuItem,
-	DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import {
-	Popover,
-	PopoverContent,
-	PopoverTrigger,
-} from "@/components/ui/popover";
-import {
-	Select,
-	SelectContent,
-	SelectItem,
-	SelectTrigger,
-	SelectValue,
-} from "@/components/ui/select";
-import { Spinner } from "@/components/ui/spinner";
-import {
 	Tooltip,
 	TooltipContent,
 	TooltipTrigger,
 } from "@/components/ui/tooltip";
-import type {
-	LogEntry,
-	LogLevel,
-} from "@/features/containers/api/get-container-logs-parsed";
-import {
-	getContainerLogsParsed,
-	getLogLevelBadgeColor,
-	streamContainerLogsParsed,
-} from "@/features/containers/api/get-container-logs-parsed";
 import { getContainers } from "@/features/containers/api/get-containers";
-import { CollapsibleJson } from "@/features/containers/components/collapsible-json";
 import {
 	formatContainerName,
 	formatCPUPercent,
@@ -79,13 +23,12 @@ import {
 	toTitleCase,
 } from "@/features/containers/components/container-utils";
 import { EnvironmentVariables } from "@/features/containers/components/environment-variables";
-import { SelectionActionBar } from "@/features/containers/components/selection-action-bar";
+import type { LogViewerHandle } from "@/features/containers/components/log-viewer/log-viewer";
+import { LogViewer } from "@/features/containers/components/log-viewer/log-viewer";
 import { Terminal } from "@/features/containers/components/terminal";
-import { useContainerLogStream } from "@/features/containers/hooks/use-container-log-stream";
 import { useContainerStats } from "@/features/containers/hooks/use-container-stats";
 import { requireAuthIfEnabled } from "@/lib/auth-guard";
-import { isJsonString } from "@/lib/json-format";
-import { escapeRegExp } from "@/lib/utils";
+
 export const Route = createFileRoute("/containers/$containerId/logs")({
 	beforeLoad: async () => {
 		await requireAuthIfEnabled();
@@ -93,63 +36,15 @@ export const Route = createFileRoute("/containers/$containerId/logs")({
 	component: ContainerLogsPage,
 });
 
-const EMPTY_SEARCH = { regex: null as RegExp | null, error: null as string | null };
-
 function ContainerLogsPage() {
 	const { containerId: encodedContainerId } = Route.useParams();
 	const navigate = useNavigate();
 	const queryClient = useQueryClient();
 
-	const [logLines, setLogLines] = useState(100);
-	const [searchText, setSearchText] = useState("");
-	const [useRegex, setUseRegex] = useState(false);
-	const [excludeMatches, setExcludeMatches] = useState(false);
-	const [selectedLevels, setSelectedLevels] = useState<Set<LogLevel>>(
-		new Set(),
-	);
-	const [showTimestamps, setShowTimestamps] = useState(true);
-	const [autoScroll, setAutoScroll] = useState(true);
-	const [wrapText, setWrapText] = useState(false);
-	const [showFilters, setShowFilters] = useState(false);
 	const [showLabels, setShowLabels] = useState(false);
 	const [showEnvVariables, setShowEnvVariables] = useState(false);
 	const [showTerminal, setShowTerminal] = useState(false);
-	const [currentMatchIndex, setCurrentMatchIndex] = useState(0);
-	const [selectedIndices, setSelectedIndices] = useState<Set<number>>(
-		new Set(),
-	);
-	const [pinnedLogIndices, setPinnedLogIndices] = useState<Set<number>>(
-		new Set(),
-	);
-	const [currentPinnedIndex, setCurrentPinnedIndex] = useState(0);
-	const [expandedJsonRows, setExpandedJsonRows] = useState<Set<number>>(
-		new Set(),
-	);
-	const [lastClickedIndex, setLastClickedIndex] = useState<number | null>(null);
-	const parentRef = useRef<HTMLDivElement>(null);
-	const searchInputRef = useRef<HTMLInputElement>(null);
-	const autoScrollRef = useRef(autoScroll);
-	const logLinesInputId = useId();
-
-	const searchParsed = useMemo(() => {
-		if (!searchText || !useRegex) return EMPTY_SEARCH;
-		try {
-			return { regex: new RegExp(searchText, "gi"), error: null as string | null };
-		} catch {
-			return { regex: null, error: "Invalid regex" };
-		}
-	}, [searchText, useRegex]);
-
-	// Pre-compiled regex for plain-text highlighting (avoids per-row compilation)
-	const plainTextSplitRegex = useMemo(() => {
-		if (!searchText || useRegex) return null;
-		return new RegExp(`(${escapeRegExp(searchText)})`, "gi");
-	}, [searchText, useRegex]);
-
-	// Keep ref in sync with state to avoid stale closures in setTimeout
-	useEffect(() => {
-		autoScrollRef.current = autoScroll;
-	}, [autoScroll]);
+	const logViewerRef = useRef<LogViewerHandle>(null);
 
 	// Decode the URL parameter (could be name or ID)
 	const containerIdentifier = decodeURIComponent(encodedContainerId);
@@ -183,669 +78,8 @@ function ContainerLogsPage() {
 
 	const handleContainerRecreated = async (_newContainerId: string) => {
 		await queryClient.invalidateQueries({ queryKey: ["containers"] });
-		if (isStreaming) {
-			stopStreaming();
-			await new Promise((resolve) => setTimeout(resolve, 100));
-			void startStreaming();
-		} else {
-			// If not streaming, just refetch logs
-			await fetchLogs();
-		}
+		await logViewerRef.current?.refreshAfterRecreate();
 	};
-
-	const scrollToBottom = useCallback((behavior: ScrollBehavior = "auto") => {
-		const containerEl = parentRef.current;
-		if (autoScrollRef.current && containerEl) {
-			if (behavior === "smooth") {
-				containerEl.scrollTo({
-					top: containerEl.scrollHeight,
-					behavior: "smooth",
-				});
-				return;
-			}
-			containerEl.scrollTop = containerEl.scrollHeight;
-		}
-	}, []);
-	const {
-		animatedRange,
-		bufferedCount,
-		droppedCount,
-		fetchLogs,
-		isLoadingLogs,
-		isStreamPaused,
-		isStreaming,
-		logs,
-		startStreaming,
-		stopStreaming,
-		togglePauseStreaming,
-		toggleStreaming,
-	} = useContainerLogStream<LogEntry>({
-		containerId: actualContainerId,
-		host: container?.host,
-		tail: logLines,
-		getLogs: getContainerLogsParsed,
-		streamLogs: streamContainerLogsParsed,
-		scrollToBottom,
-		onResetState: () => {
-			setPinnedLogIndices(new Set());
-			setCurrentPinnedIndex(0);
-		},
-		onFetchError: (error) => {
-			toast.error(`Failed to fetch logs: ${error.message}`);
-		},
-		onStreamError: (error) => {
-			toast.error(`Failed to start streaming: ${error.message}`);
-		},
-	});
-
-	// Pins are stored as indices into `logs`; when the ring buffer drops the
-	// oldest entries those indices shift down by the dropped amount.
-	const prevDroppedCountRef = useRef(0);
-	useEffect(() => {
-		const delta = droppedCount - prevDroppedCountRef.current;
-		prevDroppedCountRef.current = droppedCount;
-		if (delta <= 0) return;
-		setPinnedLogIndices((prev) => {
-			if (prev.size === 0) return prev;
-			const next = new Set<number>();
-			prev.forEach((index) => {
-				if (index - delta >= 0) {
-					next.add(index - delta);
-				}
-			});
-			return next;
-		});
-	}, [droppedCount]);
-
-	const handleRefresh = () => {
-		if (!isStreaming) {
-			fetchLogs();
-		}
-	};
-
-	useEffect(() => {
-		if (!isStreaming) {
-			fetchLogs();
-		}
-		return () => {
-			stopStreaming();
-		};
-	}, [isStreaming, fetchLogs, stopStreaming]);
-
-	const handleLogLinesChange = (value: string) => {
-		const num = parseInt(value, 10);
-		if (!Number.isNaN(num) && num > 0) {
-			setLogLines(num);
-		}
-	};
-
-	const toggleLogLevel = (level: LogLevel) => {
-		setSelectedLevels((prev) => {
-			const newSet = new Set(prev);
-			if (newSet.has(level)) {
-				newSet.delete(level);
-			} else {
-				newSet.add(level);
-			}
-			return newSet;
-		});
-	};
-
-	const handleCopyLog = (entry: LogEntry) => {
-		const text = entry.message || entry.raw || "";
-		navigator.clipboard
-			.writeText(text)
-			.then(() => {
-				toast.success("Log entry copied to clipboard");
-			})
-			.catch(() => {
-				toast.error("Failed to copy to clipboard");
-			});
-	};
-
-	const clearSelection = useCallback(() => {
-		setSelectedIndices(new Set());
-		setLastClickedIndex(null);
-	}, []);
-	const activeToggleButtonClass =
-		"h-8 text-xs data-[active=true]:bg-muted data-[active=true]:text-foreground data-[active=true]:border-border data-[active=true]:ring-1 data-[active=true]:ring-primary/30 dark:data-[active=true]:bg-primary/15 dark:data-[active=true]:ring-primary/50 dark:data-[active=true]:border-primary/30";
-
-	const toggleJsonExpanded = useCallback((index: number) => {
-		setExpandedJsonRows((prev) => {
-			const next = new Set(prev);
-			if (next.has(index)) next.delete(index);
-			else next.add(index);
-			return next;
-		});
-	}, []);
-
-	const handleLogClick = useCallback(
-		(index: number, event: React.MouseEvent) => {
-			if (event.shiftKey && lastClickedIndex !== null) {
-				// Shift-click: range selection
-				const start = Math.min(lastClickedIndex, index);
-				const end = Math.max(lastClickedIndex, index);
-				const newSelected = new Set<number>();
-				for (let i = start; i <= end; i++) {
-					newSelected.add(i);
-				}
-				setSelectedIndices(newSelected);
-			} else {
-				// Regular click: toggle single selection and set anchor
-				setSelectedIndices((prev) => {
-					const newSet = new Set(prev);
-					if (newSet.has(index)) {
-						newSet.delete(index);
-					} else {
-						newSet.add(index);
-					}
-					return newSet;
-				});
-				setLastClickedIndex(index);
-			}
-		},
-		[lastClickedIndex],
-	);
-
-	const handleDownloadLogs = (format: "json" | "txt") => {
-		if (filteredLogs.length === 0) {
-			toast.error("No logs to download");
-			return;
-		}
-
-		const containerName = (container?.names?.[0] || "container")
-			.replace(/^\//, "")
-			.replace(/[/\\:*?"<>|]/g, "-");
-		const timestamp = new Date()
-			.toISOString()
-			.replace(/:/g, "-")
-			.replace(/\..+/, "");
-		const filename = `${containerName}-logs-${timestamp}.${format}`;
-		let content: string;
-		let mimeType: string;
-
-		if (format === "json") {
-			content = JSON.stringify(filteredLogs, null, 2);
-			mimeType = "application/json";
-		} else {
-			content = filteredLogs
-				.map((entry) => {
-					const timestamp = entry.timestamp
-						? new Date(entry.timestamp).toISOString()
-						: "";
-					const level = entry.level || "UNKNOWN";
-					const message = entry.message || entry.raw || "";
-					return `[${timestamp}] [${level}] ${message}`;
-				})
-				.join("\n");
-			mimeType = "text/plain";
-		}
-
-		const blob = new Blob([content], { type: mimeType });
-		const url = URL.createObjectURL(blob);
-		const a = document.createElement("a");
-		a.href = url;
-		a.download = filename;
-		document.body.appendChild(a);
-		a.click();
-		document.body.removeChild(a);
-		URL.revokeObjectURL(url);
-		toast.success(`Logs downloaded as ${format.toUpperCase()}`);
-	};
-
-	// Filter logs by level and optionally exclude search matches.
-	const filteredLogItems = useMemo(() => {
-		const searchLower = searchText.toLowerCase();
-
-		const items = logs
-			.map((entry, originalIndex) => ({ entry, originalIndex }))
-			.filter(({ entry }) => {
-				if (selectedLevels.size > 0 && entry.level) {
-					if (!selectedLevels.has(entry.level)) {
-						return false;
-					}
-				}
-				if (excludeMatches && searchText) {
-					if (useRegex && !searchParsed.regex) {
-						// Invalid regex — skip filtering
-					} else {
-						const message = entry.message || entry.raw || "";
-						if (useRegex && searchParsed.regex) {
-							searchParsed.regex.lastIndex = 0;
-							if (searchParsed.regex.test(message)) return false;
-						} else {
-							if (message.toLowerCase().includes(searchLower)) return false;
-						}
-					}
-				}
-				return true;
-			});
-
-		return items;
-	}, [logs, selectedLevels, excludeMatches, searchText, useRegex, searchParsed]);
-
-	const filteredLogs = useMemo(
-		() => filteredLogItems.map((item) => item.entry),
-		[filteredLogItems],
-	);
-
-	const filteredToOriginalIndex = useMemo(
-		() => filteredLogItems.map((item) => item.originalIndex),
-		[filteredLogItems],
-	);
-
-	const originalToFilteredIndex = useMemo(() => {
-		const map = new Map<number, number>();
-		filteredToOriginalIndex.forEach((originalIndex, filteredIndex) => {
-			map.set(originalIndex, filteredIndex);
-		});
-		return map;
-	}, [filteredToOriginalIndex]);
-
-	const pinnedFilteredIndices = useMemo(() => {
-		const next = new Set<number>();
-		filteredToOriginalIndex.forEach((originalIndex, filteredIndex) => {
-			if (pinnedLogIndices.has(originalIndex)) {
-				next.add(filteredIndex);
-			}
-		});
-		return next;
-	}, [filteredToOriginalIndex, pinnedLogIndices]);
-
-	const selectedOriginalIndices = useMemo(() => {
-		return Array.from(selectedIndices)
-			.map((index) => filteredToOriginalIndex[index] ?? -1)
-			.filter((index) => index >= 0);
-	}, [selectedIndices, filteredToOriginalIndex]);
-
-	const allSelectedArePinned = useMemo(() => {
-		return (
-			selectedOriginalIndices.length > 0 &&
-			selectedOriginalIndices.every((index) => pinnedLogIndices.has(index))
-		);
-	}, [selectedOriginalIndices, pinnedLogIndices]);
-
-	const sortedPinnedIndices = useMemo(() => {
-		return Array.from(pinnedLogIndices).sort((a, b) => a - b);
-	}, [pinnedLogIndices]);
-
-	const handleTogglePinSelected = useCallback(() => {
-		if (selectedOriginalIndices.length === 0) return;
-
-		setPinnedLogIndices((prev) => {
-			const next = new Set(prev);
-			if (allSelectedArePinned) {
-				selectedOriginalIndices.forEach((index) => {
-					next.delete(index);
-				});
-			} else {
-				selectedOriginalIndices.forEach((index) => {
-					next.add(index);
-				});
-			}
-			return next;
-		});
-
-		toast.success(
-			allSelectedArePinned
-				? `${selectedOriginalIndices.length} ${selectedOriginalIndices.length === 1 ? "line" : "lines"} unpinned`
-				: `${selectedOriginalIndices.length} ${selectedOriginalIndices.length === 1 ? "line" : "lines"} pinned`,
-		);
-	}, [selectedOriginalIndices, allSelectedArePinned]);
-
-	const handleCopySelected = useCallback(() => {
-		if (selectedIndices.size === 0) return;
-
-		const sortedIndices = Array.from(selectedIndices).sort((a, b) => a - b);
-		// Filter out any invalid indices as a safety measure
-		const validIndices = sortedIndices.filter(
-			(idx) => idx >= 0 && idx < filteredLogs.length,
-		);
-		if (validIndices.length === 0) {
-			clearSelection();
-			return;
-		}
-
-		const selectedLogs = validIndices.map((idx) => filteredLogs[idx]);
-
-		const content = selectedLogs
-			.map((entry) => {
-				const timestamp = entry.timestamp
-					? new Date(entry.timestamp).toISOString()
-					: "";
-				const level = entry.level || "UNKNOWN";
-				const message = entry.message || entry.raw || "";
-				return `[${timestamp}] [${level}] ${message}`;
-			})
-			.join("\n");
-
-		navigator.clipboard
-			.writeText(content)
-			.then(() => {
-				toast.success(`${validIndices.length} log entries copied to clipboard`);
-			})
-			.catch(() => {
-				toast.error("Failed to copy to clipboard");
-			});
-	}, [selectedIndices, filteredLogs, clearSelection]);
-
-	// Find all matching log indices for search navigation
-	const searchMatches = useMemo(() => {
-		if (!searchText) return [];
-		if (useRegex && searchParsed.error) return [];
-		const matches: number[] = [];
-		filteredLogs.forEach((entry, index) => {
-			const message = entry.message || entry.raw || "";
-			if (useRegex && searchParsed.regex) {
-				searchParsed.regex.lastIndex = 0;
-				if (searchParsed.regex.test(message)) {
-					matches.push(index);
-				}
-			} else {
-				if (message.toLowerCase().includes(searchText.toLowerCase())) {
-					matches.push(index);
-				}
-			}
-		});
-		return matches;
-	}, [filteredLogs, searchText, useRegex, searchParsed]);
-
-	// Set view of searchMatches for O(1) per-row lookups during rendering
-	const searchMatchSet = useMemo(() => new Set(searchMatches), [searchMatches]);
-
-	// Reset current match index when search changes
-	// biome-ignore lint/correctness/useExhaustiveDependencies: intentionally reset when searchText or useRegex changes
-	useEffect(() => {
-		setCurrentMatchIndex(0);
-	}, [searchText, useRegex]);
-
-	// Clear selection when filters or search settings change
-	// biome-ignore lint/correctness/useExhaustiveDependencies: intentionally clear selection on data changes
-	useEffect(() => {
-		clearSelection();
-		setExpandedJsonRows(new Set());
-	}, [searchText, excludeMatches, selectedLevels, useRegex]);
-
-	useEffect(() => {
-		if (sortedPinnedIndices.length === 0) {
-			setCurrentPinnedIndex(0);
-		} else if (currentPinnedIndex >= sortedPinnedIndices.length) {
-			setCurrentPinnedIndex(sortedPinnedIndices.length - 1);
-		}
-	}, [sortedPinnedIndices, currentPinnedIndex]);
-
-	const availableLogLevels = useMemo(() => {
-		const levels = new Set<LogLevel>();
-		logs.forEach((entry) => {
-			if (entry.level) {
-				levels.add(entry.level);
-			}
-		});
-		return Array.from(levels).sort();
-	}, [logs]);
-
-	// Virtualization setup (must be before navigation functions)
-	const rowVirtualizer = useVirtualizer({
-		count: filteredLogs.length,
-		getScrollElement: () => parentRef.current,
-		estimateSize: () => (wrapText ? 60 : 36),
-		overscan: 5,
-	});
-
-	// Navigate to previous match
-	const goToPreviousMatch = useCallback(() => {
-		if (searchMatches.length === 0) return;
-		const newIndex =
-			currentMatchIndex > 0 ? currentMatchIndex - 1 : searchMatches.length - 1;
-		setCurrentMatchIndex(newIndex);
-		rowVirtualizer.scrollToIndex(searchMatches[newIndex], { align: "center" });
-	}, [searchMatches, currentMatchIndex, rowVirtualizer]);
-
-	// Navigate to next match
-	const goToNextMatch = useCallback(() => {
-		if (searchMatches.length === 0) return;
-		const newIndex =
-			currentMatchIndex < searchMatches.length - 1 ? currentMatchIndex + 1 : 0;
-		setCurrentMatchIndex(newIndex);
-		rowVirtualizer.scrollToIndex(searchMatches[newIndex], { align: "center" });
-	}, [searchMatches, currentMatchIndex, rowVirtualizer]);
-
-	const goToPinnedByOffset = useCallback(
-		(offset: 1 | -1) => {
-			if (sortedPinnedIndices.length === 0) return;
-
-			const newPinnedIndex =
-				(currentPinnedIndex + offset + sortedPinnedIndices.length) %
-				sortedPinnedIndices.length;
-
-			setCurrentPinnedIndex(newPinnedIndex);
-			const targetOriginalIndex = sortedPinnedIndices[newPinnedIndex];
-			const targetFilteredIndex =
-				originalToFilteredIndex.get(targetOriginalIndex) ?? -1;
-
-			if (targetFilteredIndex === -1) {
-				toast.info("Pinned line is hidden by current filters");
-				return;
-			}
-
-			setSelectedIndices(new Set([targetFilteredIndex]));
-			setLastClickedIndex(targetFilteredIndex);
-			rowVirtualizer.scrollToIndex(targetFilteredIndex, { align: "center" });
-		},
-		[
-			sortedPinnedIndices,
-			currentPinnedIndex,
-			originalToFilteredIndex,
-			rowVirtualizer,
-		],
-	);
-
-	const goToAdjacentLogLine = useCallback(
-		(direction: 1 | -1) => {
-			if (filteredLogs.length === 0) return;
-			const selected = Array.from(selectedIndices);
-			const fallbackIndex =
-				selected.length > 0
-					? direction > 0
-						? Math.max(...selected)
-						: Math.min(...selected)
-					: direction > 0
-						? -1
-						: filteredLogs.length;
-			const baseIndex = lastClickedIndex ?? fallbackIndex;
-			const nextIndex = Math.min(
-				filteredLogs.length - 1,
-				Math.max(0, baseIndex + direction),
-			);
-			if (nextIndex === baseIndex) return;
-
-			setSelectedIndices(new Set([nextIndex]));
-			setLastClickedIndex(nextIndex);
-			rowVirtualizer.scrollToIndex(nextIndex, { align: "center" });
-		},
-		[filteredLogs.length, lastClickedIndex, rowVirtualizer, selectedIndices],
-	);
-
-	const extendSelectionByLine = useCallback(
-		(direction: 1 | -1) => {
-			if (filteredLogs.length === 0) return;
-
-			const selected = Array.from(selectedIndices);
-			const minSelected = selected.length > 0 ? Math.min(...selected) : null;
-			const maxSelected = selected.length > 0 ? Math.max(...selected) : null;
-
-			let anchorIndex: number;
-			if (lastClickedIndex !== null) {
-				anchorIndex = lastClickedIndex;
-			} else {
-				if (selected.length > 0) {
-					anchorIndex =
-						direction > 0 ? (minSelected as number) : (maxSelected as number);
-				} else {
-					anchorIndex = direction > 0 ? 0 : filteredLogs.length - 1;
-				}
-				setLastClickedIndex(anchorIndex);
-			}
-
-			const activeIndex =
-				selected.length > 0
-					? direction > 0
-						? (maxSelected as number)
-						: (minSelected as number)
-					: anchorIndex;
-			const targetIndex = Math.min(
-				filteredLogs.length - 1,
-				Math.max(0, activeIndex + direction),
-			);
-			if (targetIndex === activeIndex) return;
-
-			const rangeStart = Math.min(anchorIndex, targetIndex);
-			const rangeEnd = Math.max(anchorIndex, targetIndex);
-			const nextSelected = new Set<number>();
-			for (let i = rangeStart; i <= rangeEnd; i++) {
-				nextSelected.add(i);
-			}
-			setSelectedIndices(nextSelected);
-			rowVirtualizer.scrollToIndex(targetIndex, { align: "center" });
-		},
-		[filteredLogs.length, lastClickedIndex, rowVirtualizer, selectedIndices],
-	);
-
-	const focusSearchInput = useCallback(() => {
-		searchInputRef.current?.focus();
-		searchInputRef.current?.select();
-	}, []);
-
-	useEffect(() => {
-		const handleGlobalKeyDown = (event: KeyboardEvent) => {
-			if (
-				event.defaultPrevented ||
-				event.metaKey ||
-				event.ctrlKey ||
-				event.altKey
-			)
-				return;
-			const lowerKey = event.key.toLowerCase();
-
-			const target = event.target;
-			if (
-				target instanceof HTMLElement &&
-				(target.isContentEditable ||
-					target.closest(
-						"input, textarea, select, [contenteditable='true'], [role='textbox']",
-					))
-			) {
-				return;
-			}
-
-			if (event.key === "/") {
-				event.preventDefault();
-				focusSearchInput();
-				return;
-			}
-
-			if (lowerKey === "j" || event.key === "ArrowDown") {
-				event.preventDefault();
-				if (event.shiftKey) {
-					extendSelectionByLine(1);
-				} else {
-					goToAdjacentLogLine(1);
-				}
-				return;
-			}
-
-			if (lowerKey === "k" || event.key === "ArrowUp") {
-				event.preventDefault();
-				if (event.shiftKey) {
-					extendSelectionByLine(-1);
-				} else {
-					goToAdjacentLogLine(-1);
-				}
-				return;
-			}
-
-			if (lowerKey === "n") {
-				event.preventDefault();
-				if (event.shiftKey) {
-					goToPreviousMatch();
-				} else {
-					goToNextMatch();
-				}
-				return;
-			}
-
-			if (lowerKey === "p") {
-				event.preventDefault();
-				if (event.shiftKey) {
-					goToPinnedByOffset(-1);
-				} else {
-					goToPinnedByOffset(1);
-				}
-			}
-		};
-
-		window.addEventListener("keydown", handleGlobalKeyDown);
-		return () => {
-			window.removeEventListener("keydown", handleGlobalKeyDown);
-		};
-	}, [
-		focusSearchInput,
-		goToAdjacentLogLine,
-		extendSelectionByLine,
-		goToNextMatch,
-		goToPinnedByOffset,
-		goToPreviousMatch,
-	]);
-
-	const highlightSearchText = useCallback(
-		(text: string, isCurrentMatch: boolean): React.ReactNode => {
-			if (!searchText || !text) return text;
-
-			if (useRegex && searchParsed.regex) {
-				searchParsed.regex.lastIndex = 0;
-				const parts: React.ReactNode[] = [];
-				let lastIndex = 0;
-				let key = 0;
-				for (const match of text.matchAll(searchParsed.regex)) {
-					if (match.index === undefined) continue;
-					// Guard against zero-length matches to avoid infinite loops
-					if (match[0].length === 0) continue;
-					if (match.index > lastIndex) {
-						parts.push(<span key={key++}>{text.slice(lastIndex, match.index)}</span>);
-					}
-					parts.push(
-						<mark
-							key={key++}
-							className={`px-0.5 rounded ${isCurrentMatch ? "bg-yellow-400 dark:bg-yellow-500" : "bg-yellow-200 dark:bg-yellow-700"}`}
-						>
-							{match[0]}
-						</mark>
-					);
-					lastIndex = match.index + match[0].length;
-				}
-				if (lastIndex < text.length) {
-					parts.push(<span key={key++}>{text.slice(lastIndex)}</span>);
-				}
-				return parts.length > 0 ? parts : text;
-			}
-
-			if (!plainTextSplitRegex) return text;
-			const splitParts = text.split(plainTextSplitRegex);
-			return splitParts.map((part, i) =>
-				part.toLowerCase() === searchText.toLowerCase() ? (
-					<mark
-						key={`${i}-${part}`}
-						className={`px-0.5 rounded ${isCurrentMatch ? "bg-yellow-400 dark:bg-yellow-500" : "bg-yellow-200 dark:bg-yellow-700"}`}
-					>
-						{part}
-					</mark>
-				) : (
-					<span key={`${i}-${part}`}>{part}</span>
-				),
-			);
-		},
-		[searchText, useRegex, searchParsed, plainTextSplitRegex],
-	);
 
 	return (
 		<div className="min-h-screen bg-background">
@@ -1094,505 +328,13 @@ function ContainerLogsPage() {
 					)}
 
 					{/* Logs Card */}
-					<Card>
-						<CardHeader>
-							<div className="space-y-3">
-								{/* Row 1: Search + Lines + Stream/Refresh */}
-								<div className="flex items-center gap-2">
-									<CardTitle className="text-base shrink-0">
-										Logs
-										{filteredLogs.length !== logs.length && (
-											<span className="ml-2 text-xs text-muted-foreground font-normal">
-												({filteredLogs.length} of {logs.length})
-											</span>
-										)}
-									</CardTitle>
-
-									<div className="relative flex-1 min-w-[140px]">
-										<SearchIcon className="absolute left-2.5 top-1/2 -translate-y-1/2 size-3.5 text-muted-foreground" />
-										<Input
-											ref={searchInputRef}
-											placeholder="Search logs..."
-											value={searchText}
-											onChange={(e) => setSearchText(e.target.value)}
-											className={`pl-8 h-8 text-xs ${useRegex && searchParsed.error ? "border-red-500 focus-visible:ring-red-500" : ""}`}
-										/>
-									</div>
-									<Button
-										variant={useRegex ? "secondary" : "ghost"}
-										size="sm"
-										onClick={() => setUseRegex(!useRegex)}
-										className="h-8 w-8 p-0 font-mono text-xs shrink-0"
-										title={useRegex ? "Switch to plain text search" : "Switch to regex search"}
-									>
-										.*
-									</Button>
-									{searchText && !excludeMatches && (
-										<div className="flex items-center gap-0.5 shrink-0">
-											<span className="text-xs tabular-nums text-muted-foreground whitespace-nowrap px-1">
-												{searchMatches.length > 0
-													? `${currentMatchIndex + 1} of ${searchMatches.length}`
-													: "No matches"}
-											</span>
-											<Button
-												variant="outline"
-												size="sm"
-												onClick={goToPreviousMatch}
-												disabled={searchMatches.length === 0}
-												className="h-8 w-8 p-0"
-											>
-												<ChevronLeftIcon className="size-3.5" />
-											</Button>
-											<Button
-												variant="outline"
-												size="sm"
-												onClick={goToNextMatch}
-												disabled={searchMatches.length === 0}
-												className="h-8 w-8 p-0"
-											>
-												<ChevronRightIcon className="size-3.5" />
-											</Button>
-										</div>
-									)}
-
-									{sortedPinnedIndices.length > 0 && (
-										<div className="flex items-center gap-0.5 shrink-0">
-											<span className="text-xs tabular-nums text-muted-foreground whitespace-nowrap px-1">
-												{`${Math.min(currentPinnedIndex + 1, sortedPinnedIndices.length)} of ${sortedPinnedIndices.length} pinned`}
-											</span>
-											<Button
-												variant="outline"
-												size="sm"
-												onClick={() => goToPinnedByOffset(-1)}
-												className="h-8 w-8 p-0"
-											>
-												<ChevronLeftIcon className="size-3.5" />
-											</Button>
-											<Button
-												variant="outline"
-												size="sm"
-												onClick={() => goToPinnedByOffset(1)}
-												className="h-8 w-8 p-0"
-											>
-												<ChevronRightIcon className="size-3.5" />
-											</Button>
-										</div>
-									)}
-
-									<div className="flex items-center gap-2 shrink-0">
-										<Label
-											htmlFor={logLinesInputId}
-											className="text-xs text-muted-foreground"
-										>
-											Lines
-										</Label>
-										<Input
-											id={logLinesInputId}
-											type="text"
-											inputMode="numeric"
-											pattern="[0-9]*"
-											value={logLines}
-											onChange={(e) => handleLogLinesChange(e.target.value)}
-											disabled={isStreaming}
-											className="h-8 w-20 text-xs"
-										/>
-									</div>
-									<Button
-										variant="outline"
-										size="sm"
-										data-active={isStreaming}
-										onClick={toggleStreaming}
-										disabled={isLoadingLogs && !isStreaming}
-										className={`shrink-0 ${activeToggleButtonClass}`}
-									>
-										{isStreaming ? (
-											<>
-												<SquareIcon className="mr-2 size-4" />
-												Stop
-											</>
-										) : (
-											<>
-												<PlayIcon className="mr-2 size-4" />
-												Stream
-											</>
-										)}
-									</Button>
-									<Button
-										variant="outline"
-										size="sm"
-										data-active={isStreamPaused}
-										onClick={togglePauseStreaming}
-										disabled={!isStreaming}
-										className={`shrink-0 ${activeToggleButtonClass}`}
-									>
-										{isStreamPaused ? (
-											<>
-												<PlayIcon className="mr-2 size-4" />
-												Resume
-												{bufferedCount > 0 && (
-													<span className="ml-1 text-[10px] tabular-nums">
-														({bufferedCount})
-													</span>
-												)}
-											</>
-										) : (
-											<>
-												<PauseIcon className="mr-2 size-4" />
-												Pause
-											</>
-										)}
-									</Button>
-									<Button
-										variant="outline"
-										size="sm"
-										onClick={handleRefresh}
-										disabled={isStreaming || isLoadingLogs}
-										className="shrink-0"
-									>
-										<RefreshCcwIcon className="size-4" />
-									</Button>
-								</div>
-								<p className="text-[11px] text-muted-foreground">
-									Shortcuts: <kbd className="font-mono">/</kbd> search,{" "}
-									<kbd className="font-mono">j</kbd>/
-									<kbd className="font-mono">k</kbd> lines,{" "}
-									<kbd className="font-mono">n</kbd>/
-									<kbd className="font-mono">N</kbd> matches,{" "}
-									<kbd className="font-mono">p</kbd>/
-									<kbd className="font-mono">P</kbd> pins
-								</p>
-								{/* Row 2: Options bar */}
-								<div className="flex flex-wrap items-center gap-2">
-									{searchText && (
-										<Select
-											value={excludeMatches ? "exclude" : "highlight"}
-											onValueChange={(v) => setExcludeMatches(v === "exclude")}
-										>
-											<SelectTrigger size="sm" className="w-[160px] text-xs">
-												<SelectValue />
-											</SelectTrigger>
-											<SelectContent>
-												<SelectItem value="highlight">
-													Highlight matches
-												</SelectItem>
-												<SelectItem value="exclude">Exclude matches</SelectItem>
-											</SelectContent>
-										</Select>
-									)}
-
-									<Popover open={showFilters} onOpenChange={setShowFilters}>
-										<PopoverTrigger asChild>
-											<Button
-												variant="outline"
-												size="sm"
-												className="h-8 text-xs"
-											>
-												Log level
-												{selectedLevels.size > 0 && (
-													<Badge
-														variant="secondary"
-														className="ml-1.5 px-1 py-0 h-4 text-[10px] leading-none"
-													>
-														{selectedLevels.size}
-													</Badge>
-												)}
-												<ChevronDownIcon className="ml-1 size-3.5 opacity-50" />
-											</Button>
-										</PopoverTrigger>
-										<PopoverContent align="start" className="w-56">
-											<div className="space-y-3">
-												<div>
-													<h4 className="text-sm font-medium mb-2">
-														Log Levels
-													</h4>
-													<div className="space-y-2">
-														{availableLogLevels.length === 0 ? (
-															<p className="text-xs text-muted-foreground">
-																No log levels available
-															</p>
-														) : (
-															availableLogLevels.map((level) => (
-																<label
-																	key={level}
-																	className="flex items-center gap-2 cursor-pointer"
-																>
-																	<button
-																		type="button"
-																		onClick={() => toggleLogLevel(level)}
-																		className={`size-4 rounded border flex items-center justify-center ${
-																			selectedLevels.has(level)
-																				? "bg-primary border-primary"
-																				: "border-input"
-																		}`}
-																	>
-																		{selectedLevels.has(level) && (
-																			<CheckIcon className="size-3 text-primary-foreground" />
-																		)}
-																	</button>
-																	<Badge
-																		variant="outline"
-																		className={`text-xs ${getLogLevelBadgeColor(level)}`}
-																	>
-																		{level}
-																	</Badge>
-																</label>
-															))
-														)}
-													</div>
-												</div>
-												{selectedLevels.size > 0 && (
-													<Button
-														variant="outline"
-														size="sm"
-														onClick={() => setSelectedLevels(new Set())}
-														className="w-full"
-													>
-														Clear Filters
-													</Button>
-												)}
-											</div>
-										</PopoverContent>
-									</Popover>
-
-									<Button
-										variant="outline"
-										size="sm"
-										data-active={showTimestamps}
-										onClick={() => setShowTimestamps(!showTimestamps)}
-										className={activeToggleButtonClass}
-									>
-										Timestamps
-									</Button>
-
-									<Button
-										variant="outline"
-										size="sm"
-										data-active={wrapText}
-										onClick={() => setWrapText(!wrapText)}
-										className={activeToggleButtonClass}
-									>
-										Wrap
-									</Button>
-
-									<Button
-										variant="outline"
-										size="sm"
-										data-active={autoScroll}
-										onClick={() => setAutoScroll(!autoScroll)}
-										className={activeToggleButtonClass}
-									>
-										{autoScroll ? (
-											<ArrowDownToLineIcon className="mr-1.5 size-3.5" />
-										) : (
-											<ArrowDownIcon className="mr-1.5 size-3.5" />
-										)}
-										Auto-scroll
-									</Button>
-
-									<div className="flex-1" />
-
-									<DropdownMenu>
-										<DropdownMenuTrigger asChild>
-											<Button
-												variant="outline"
-												size="sm"
-												className="h-8 text-xs"
-											>
-												<DownloadIcon className="mr-1.5 size-3.5" />
-												Download
-											</Button>
-										</DropdownMenuTrigger>
-										<DropdownMenuContent align="end">
-											<DropdownMenuItem
-												onClick={() => handleDownloadLogs("json")}
-											>
-												Download as JSON
-											</DropdownMenuItem>
-											<DropdownMenuItem
-												onClick={() => handleDownloadLogs("txt")}
-											>
-												Download as TXT
-											</DropdownMenuItem>
-										</DropdownMenuContent>
-									</DropdownMenu>
-								</div>
-							</div>
-						</CardHeader>
-						<CardContent className="p-0 relative">
-							{/* Selection Action Bar - sticky at top of logs */}
-							<SelectionActionBar
-								selectedCount={selectedIndices.size}
-								onCopy={handleCopySelected}
-								onTogglePin={handleTogglePinSelected}
-								pinActionLabel={allSelectedArePinned ? "unpin" : "pin"}
-								onClear={clearSelection}
-							/>
-							<div
-								ref={parentRef}
-								className="h-[calc(100vh-400px)] min-h-[400px] w-full overflow-auto"
-							>
-								{isLoadingLogs && logs.length === 0 ? (
-									<div className="flex items-center justify-center py-8 text-muted-foreground">
-										<Spinner className="mr-2 size-4" />
-										Loading logs...
-									</div>
-								) : logs.length === 0 ? (
-									<div className="flex items-center justify-center py-8 text-muted-foreground text-sm">
-										No logs available
-									</div>
-								) : filteredLogs.length === 0 ? (
-									<div className="flex items-center justify-center py-8 text-muted-foreground text-sm">
-										No logs match the current filters
-									</div>
-								) : (
-									<div
-										style={{
-											height: `${rowVirtualizer.getTotalSize()}px`,
-											width: "100%",
-											position: "relative",
-										}}
-										className={`font-mono text-xs ${wrapText ? "" : "w-fit min-w-full"}`}
-									>
-										{rowVirtualizer.getVirtualItems().map((virtualRow) => {
-											const entry = filteredLogs[virtualRow.index];
-											const displayText = entry.message || entry.raw || "";
-											if (!displayText.trim()) return null;
-
-											const timestamp = entry.timestamp
-												? new Date(entry.timestamp)
-												: null;
-											const dateLabel = timestamp
-												? `${timestamp.toLocaleDateString("en-GB", {
-														day: "2-digit",
-														month: "2-digit",
-														year: "numeric",
-													})} ${timestamp.toLocaleTimeString("en-US", {
-														hour12: false,
-														hour: "2-digit",
-														minute: "2-digit",
-														second: "2-digit",
-													})}`
-												: "—";
-
-											// Check if this row is the current search match
-											const isCurrentMatch =
-												searchMatches.length > 0 &&
-												searchMatches[currentMatchIndex] === virtualRow.index;
-											const hasMatch = searchMatchSet.has(virtualRow.index);
-											const isSelected = selectedIndices.has(virtualRow.index);
-											const isPinned = pinnedFilteredIndices.has(
-												virtualRow.index,
-											);
-
-											return (
-												// biome-ignore lint/a11y/useSemanticElements: div required for virtual scrolling absolute positioning
-												<div
-													key={virtualRow.key}
-													data-index={virtualRow.index}
-													ref={rowVirtualizer.measureElement}
-													role="button"
-													tabIndex={0}
-													onClick={(e) => handleLogClick(virtualRow.index, e)}
-													onKeyDown={(e) => {
-														if (e.key === "Enter" || e.key === " ") {
-															e.preventDefault();
-															handleLogClick(
-																virtualRow.index,
-																e as unknown as React.MouseEvent,
-															);
-														}
-													}}
-													style={{
-														position: "absolute",
-														top: 0,
-														left: 0,
-														width: wrapText ? "100%" : "max-content",
-														minWidth: "100%",
-														transform: `translateY(${virtualRow.start}px)`,
-														cursor: "pointer",
-													}}
-													className={`group flex items-start gap-3 px-4 py-1.5 transition-all duration-150 ease-out ${
-														wrapText ? "" : "whitespace-nowrap"
-													} ${
-														isSelected && isPinned
-															? "bg-amber-200/80 dark:bg-amber-800/45 border-l-[3px] border-amber-500 shadow-[inset_0_0_0_1px_rgba(245,158,11,0.35)]"
-															: isSelected
-																? "bg-primary/[0.08] dark:bg-primary/[0.15] border-l-[3px] border-primary shadow-[inset_0_0_0_1px_rgba(var(--primary),0.1)]"
-																: isPinned
-																	? "bg-amber-100/80 dark:bg-amber-900/35 border-l-[3px] border-amber-500 hover:bg-amber-100 dark:hover:bg-amber-900/45"
-																	: isCurrentMatch
-																		? "bg-yellow-100 dark:bg-yellow-900/30 border-y-2 border-yellow-400 dark:border-yellow-600"
-																		: virtualRow.index % 2 === 0
-																			? "bg-muted/30 border-l-[3px] border-transparent hover:bg-muted/50"
-																			: "border-l-[3px] border-transparent hover:bg-muted/50"
-													} ${
-														animatedRange &&
-														virtualRow.index >= animatedRange.start &&
-														virtualRow.index <= animatedRange.end
-															? "log-stream-row-enter"
-															: ""
-													}`}
-												>
-													{showTimestamps && (
-														<span className="text-muted-foreground shrink-0 text-[11px]">
-															{dateLabel}
-														</span>
-													)}
-													<Badge
-														variant="outline"
-														className={`shrink-0 text-xs px-1.5 py-0 h-5 ${getLogLevelBadgeColor(entry.level ?? "UNKNOWN")}`}
-													>
-														{entry.level ?? "UNKNOWN"}
-													</Badge>
-													<span
-														className={`text-foreground flex-1 ${wrapText ? "break-words" : ""}`}
-													>
-														{isJsonString(entry.message ?? "") ? (
-															<CollapsibleJson
-																text={entry.message ?? ""}
-																isExpanded={expandedJsonRows.has(
-																	virtualRow.index,
-																)}
-																onToggle={() =>
-																	toggleJsonExpanded(virtualRow.index)
-																}
-																isCurrentMatch={isCurrentMatch}
-																highlightSearchText={
-																	hasMatch ? highlightSearchText : undefined
-																}
-															/>
-														) : hasMatch ? (
-															highlightSearchText(
-																displayText,
-																isCurrentMatch,
-															)
-														) : (
-															displayText
-														)}
-													</span>
-													<Tooltip>
-														<TooltipTrigger asChild>
-															<button
-																type="button"
-																onClick={(e) => {
-																	e.stopPropagation();
-																	handleCopyLog(entry);
-																}}
-																className="shrink-0 opacity-0 group-hover:opacity-100 transition-opacity p-1 hover:bg-muted rounded"
-															>
-																<CopyIcon className="size-3 text-muted-foreground" />
-															</button>
-														</TooltipTrigger>
-														<TooltipContent>Copy log entry</TooltipContent>
-													</Tooltip>
-												</div>
-											);
-										})}
-									</div>
-								)}
-							</div>
-						</CardContent>
-					</Card>
+					<LogViewer
+						ref={logViewerRef}
+						variant="page"
+						containerId={actualContainerId}
+						host={container?.host}
+						containerName={container?.names?.[0]}
+					/>
 				</div>
 			</div>
 		</div>
